split response responsibility from the client
